+++ src/main/kotlin/ris58h/webcalm/javascript/psi/JavaScriptClassStaticBlock.kt
@@ -1,0 +1,9 @@
+package ris58h.webcalm.javascript.psi
+
+import com.intellij.extapi.psi.ASTWrapperPsiElement
+import com.intellij.lang.ASTNode
+
+class JavaScriptClassStaticBlock(node: ASTNode) : ASTWrapperPsiElement(node), JavaScriptClassElement {
+    val block: JavaScriptBlock?
+        get() = this.findChildByClass(JavaScriptBlock::class.java)
+}